lagrangian/localrequirements.txt
Django==1.8.16
django-location-field==1.6.1
django-registration==2.1.2
djangorestframework==3.4.7
oauthlib==2.0.0
Pillow==3.4.2
reportlab==3.3.0
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
tweepy==3.5.0
